maze_navigator work in progress...
