+++ navigation/maze_navigator/src/ring_pickup.cpp
@@ -1,7 +1,4 @@
 #include <ros/ros.h>
-#include <move_base_msgs/MoveBaseAction.h>
-#include <actionlib/client/simple_action_client.h>
-typedef actionlib::SimpleActionClient<move_base_msgs::MoveBaseAction> MoveBaseClient;
 
 #include <tf/tf.h>
 #include <tf/transform_datatypes.h>
@@ -27,17 +24,17 @@ typedef actionlib::SimpleActionClient<move_base_msgs::MoveBaseAction> MoveBaseCl
 using namespace std;
 using namespace cv;
 
-
 /*
     Small correction relative to robot.
 */
-void constructGoal(MoveBaseClient &ac, tf::Vector3& trans, double angle_2d){
+void constructGoal(MoveBaseClient &ac, tf::Vector3 &trans, double angle_2d)
+{
     move_base_msgs::MoveBaseGoal goal;
 
     goal.target_pose.header.frame_id = "/base_link";
     goal.target_pose.header.stamp = ros::Time::now();
-  
-    ROS_INFO("Moving ROBOT -> (%f %f) to (%f %f)\n",r_pos.wy_, r_pos.wx_, r_goal.wy_, r_goal.wx_);
+
+    ROS_INFO("Moving ROBOT -> (%f %f) to (%f %f)\n", r_pos.wy_, r_pos.wx_, r_goal.wy_, r_goal.wx_);
 
     // translation
     goal.target_pose.pose.position.x = trans.x();
@@ -50,19 +47,26 @@ void constructGoal(MoveBaseClient &ac, tf::Vector3& trans, double angle_2d){
     goal.target_pose.pose.orientation.y = rot_quat_.y();
     goal.target_pose.pose.orientation.z = rot_quat_.z();
     goal.target_pose.pose.orientation.w = rot_quat_.w();
-    
+
     ROS_INFO("Sending goal (%f, %f, %f) [%f, %f, %f, %f]\n",
-        goal.target_pose.pose.position.x,
-        goal.target_pose.pose.position.y,
-        goal.target_pose.pose.position.z,
-        goal.target_pose.pose.orientation.x,
-        goal.target_pose.pose.orientation.y,
-        goal.target_pose.pose.orientation.z,
-        goal.target_pose.pose.orientation.w);
-        
+             goal.target_pose.pose.position.x,
+             goal.target_pose.pose.position.y,
+             goal.target_pose.pose.position.z,
+             goal.target_pose.pose.orientation.x,
+             goal.target_pose.pose.orientation.y,
+             goal.target_pose.pose.orientation.z,
+             goal.target_pose.pose.orientation.w);
+
     ac.sendGoal(goal);
 }
 
+/*
+    Get ring position
+*/
+void ringPoseCallback(geometry_msgs::PointStamped &ring_pose)
+{
+    ROS_INFO("RING DETECTED %f %f %f\n", ring_pose.point.x, ring_pose.point.y, ring_pose.point.z);
+}
 
 int main(int argc, char **argv)
 {
@@ -72,33 +76,17 @@ int main(int argc, char **argv)
     ros::NodeHandle n;
 
     // subscribers
-    ros::Subscriber costmap_sub;
-    ros::Subscriber simplemap_sub;
+    ros::Subscriber ringpose_sub;
 
     // publishers
-    ros::Publisher goal_pub;
     ros::Publisher vis_pub;
 
-     // frame transformer
+    // frame transformer
     tf::TransformListener listener;
     tf::StampedTransform transform;
 
-    // goal_pub = n.advertise<geometry_msgs::PoseStamped>("goal", 10);
-    // vis_pub = n.advertise<visualization_msgs::MarkerArray>("visualization_marker", 10);
+    ringpose_sub = nh.subscribe("input", 1, ringPoseCallback);
     vis_pub = n.advertise<visualization_msgs::Marker>("visualization_marker", 10);
-
-    // goal sender
-    MoveBaseClient ac("move_base", true);
-
-    //wait for the action server to come up
-    while (!ac.waitForServer(ros::Duration(5.0)))
-    {
-        ROS_INFO("Waiting for the move_base action server to come up");
-    }
-
-    RobotPose r_goal; 
-    RobotPose r_start;
-
 
     // main loop
     while (ros::ok())
